requirements.txt
requests>=2.31.0
python-docx>=1.1.2
pdfplumber>=0.11.0
openai>=1.40.0
anthropic>=0.25.0
